requirements.txt
pytest==4.2
pylint==2.2
django==2.1
djangorestframework==3.9
model-mommy==1.6
requests==2.21